Issue id #1
Добавление поддержки неименованных доменов. Переименование файла.
